Started to fix compilation issues. Fixed common.
